Version 3.1
- Fixed the .weather command not returning the results if the precipitation value has decimal point.
- Updated the documentation to match the latest version.
- Moved the .ip command into the misc module, renamed the world module to weather.

+++ src/FlawBOT/Models/WorldData.cs
@@ -133,7 +133,7 @@ namespace FlawBOT.Models
         public int Pressure { get; set; }
 
         [JsonProperty("precip", NullValueHandling = NullValueHandling.Ignore)]
-        public int Precipitation { get; set; }
+        public double Precipitation { get; set; }
 
         [JsonProperty("humidity", NullValueHandling = NullValueHandling.Ignore)]
         public int Humidity { get; set; }
